Initial commit
Frontend Web Developer Internship - Task 2: To-Do List Application

Company: Cognorise Infotech

Position: Frontend Web Developer Intern

Cognorise Infotech is a dynamic and innovative tech company dedicated to pushing the boundaries of digital solutions. With a focus on cutting-edge technologies and a passion for creativity, we strive to deliver impactful solutions that empower businesses and individuals alike.